requirements.txt
torch
numpy
Pillow
rembg
trimesh
pyrender
matplotlib
open3d
